Moves getStyleAttributeValue to styleUtils
